Simplify access to the version string from a version lock
